refactor files (changed to gradle testng)
