+++ src/test/java/by/onliner/login/BaseTest.java
@@ -1,0 +1,26 @@
+package by.onliner.login;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import java.time.Duration;
+
+public class BaseTest {
+    WebDriver driver;
+    LoginPage loginPage;
+
+    @BeforeMethod
+    public void setUp() {
+        driver = new ChromeDriver();
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(4));
+        driver.get("https://www.onliner.by/");
+        loginPage = new LoginPage(driver);
+    }
+
+    @AfterMethod
+    void tearDown() {
+        driver.quit();
+    }
+}
